refactor(build): enable umd/commonjs/es builds with TypeScript
