+++ src/umd.ts
@@ -1,0 +1,3 @@
+import * as Stardust from './index'
+
+module.exports = Stardust
