--- src/umd.tsx
@@ -1,5 +1,0 @@
-import * as Stardust from './'
-
-module.exports = {
-  ...Stardust,
-}
