+++ types/react-fela.d.ts
@@ -1,0 +1,3 @@
+declare module 'react-fela' {
+  export const FelaTheme: any
+}
